Convert If node to versioned node, add v2
